requirements.txt
alembic==1.13.1
amqp==5.3.1
aniso8601==10.0.0
annotated-types==0.7.0
attrs==24.3.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
billiard==4.2.1
black==23.12.1
blinker==1.9.0
celery==5.3.6
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.10
fake-useragent==2.0.3
flake8==7.0.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
instagrapi==2.1.3
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
pipdeptree==2.24.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pyflakes==3.2.0
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.4
pytest-cov==4.1.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
redis==5.0.1
requests==2.31.0
selenium==4.16.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.25
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
